Generalize things to prepare for PATCH
