+++ enterprise/internal/scim/user_create.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/elimity-com/scim"
 	scimerrors "github.com/elimity-com/scim/errors"
-	"github.com/sourcegraph/sourcegraph/cmd/frontend/auth"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc"
 	"github.com/sourcegraph/sourcegraph/internal/types"
@@ -16,42 +15,28 @@ import (
 
 // Create stores given attributes. Returns a resource with the attributes that are stored and a (new) unique identifier.
 func (h *UserResourceHandler) Create(r *http.Request, attributes scim.ResourceAttributes) (scim.Resource, error) {
-	// Get external ID, primary email, username, and display name
+	// Extract external ID, primary email, username, and display name from attributes to variables
 	optionalExternalID := getOptionalExternalID(attributes)
 	primaryEmail := extractPrimaryEmail(attributes)
 	if primaryEmail == "" {
 		return scim.Resource{}, scimerrors.ScimErrorBadParams([]string{"emails missing"})
 	}
-	displayName := extractDisplayName(attributes)
-
-	// Process requested username
 	requestedUsername := extractUsername(attributes)
-	normalizedUsername, err := auth.NormalizeUsername(requestedUsername)
-	if err != nil {
-		normalizedUsername, err = auth.AddRandomSuffix("")
-		if err != nil {
-			return scim.Resource{}, scimerrors.ScimErrorBadParams([]string{"invalid username"})
-		}
-	}
+	displayName := extractDisplayName(attributes)
 
 	// Make sure the username is unique, then create user with/without an external account ID
 	var user *types.User
-	err = h.db.WithTransact(r.Context(), func(tx database.DB) error {
-		_, err := tx.Users().GetByUsername(r.Context(), normalizedUsername)
-		if err == nil { // Username exists, try to add random suffix
-			normalizedUsername, err = auth.AddRandomSuffix(normalizedUsername)
-			if err != nil {
-				return scimerrors.ScimError{Status: http.StatusInternalServerError, Detail: errors.Wrap(err, "could not normalize username").Error()}
-			}
-		} else if !database.IsUserNotFoundErr(err) {
-			return scimerrors.ScimError{Status: http.StatusInternalServerError, Detail: errors.Wrap(err, "could not check if username exists").Error()}
+	err := h.db.WithTransact(r.Context(), func(tx database.DB) error {
+		uniqueUsername, err := getUniqueUsername(r.Context(), tx.Users(), requestedUsername)
+		if err != nil {
+			return err
 		}
 
 		// Create user (with or without external ID)
 		// TODO: Use NewSCIMUser instead of NewUser?
 		newUser := database.NewUser{
 			Email:           primaryEmail,
-			Username:        normalizedUsername,
+			Username:        uniqueUsername,
 			DisplayName:     displayName,
 			EmailIsVerified: true,
 		}
@@ -111,15 +96,6 @@ func extractPrimaryEmail(attributes scim.ResourceAttributes) (primaryEmail strin
 	}
 	if primaryEmail == "" && len(emails) > 0 {
 		primaryEmail = emails[0].(map[string]interface{})["value"].(string)
-	}
-	return
-}
-
-// extractUsername extracts the username from the given attributes.
-func extractUsername(attributes scim.ResourceAttributes) (username string) {
-	// TODO: Validate here?
-	if attributes["userName"] != nil {
-		username = attributes["userName"].(string)
 	}
 	return
 }
